requirements.txt
streamlit>=1.40.0
anthropic>=0.42.0
PyMuPDF>=1.25.0
python-docx>=1.1.0
pyyaml>=6.0
plotly>=5.24.0
pandas>=2.2.0
reportlab>=4.2.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
extra-streamlit-components>=0.1.60
